removed numpy helpers
